HangGlider, Car, vehicle types

+++ PojazdyLab/PojazdyLib/Interfaces/IVehicle.cs
@@ -10,11 +10,11 @@ namespace PojazdyLib.Interfejsy
     public interface IVehicle
     {
         enum State { off, on } // Stan pojazdu - czy pojazd jest uruchomiony czy zatrzymany
-        int CurrentSpeed { get; } // Aktualna prędkość pojazdu
+        double CurrentSpeed { get; } // Aktualna prędkość pojazdu
         bool? CanVahicleBeStopped { get; } // Czy można zatrzymać pojazd? 
         void Start(); // Uruchamia pojazd
         void Stop(); // Zatrzymuje pojazd
-        void IncreaseSpeed(int speed); // Zwiększa prędkość pojazdu
-        void ReduceSpeed(int speed); // Zmniejsza prędkość pojazdu
+        void IncreaseSpeed(double speed); // Zwiększa prędkość pojazdu
+        void ReduceSpeed(double speed); // Zmniejsza prędkość pojazdu
     }
 }
